qos_core: add pool_size field to PivotConfig in the Manifest
and use it in Reaper, also expand qos_client with the cli argument for it

+++ src/qos_core/src/client.rs
@@ -76,7 +76,7 @@ impl SocketClient {
 	/// Expands the underlying `AsyncPool` to given `pool_size`
 	pub async fn expand_to(
 		&mut self,
-		pool_size: u32,
+		pool_size: u8,
 	) -> Result<(), ClientError> {
 		self.pool.write().await.expand_to(pool_size)?;
